backend/requirements.txt
fastapi
uvicorn
playwright
supabase
python-dotenv
pydantic
httpx
beautifulsoup4
google-generativeai
google-generativeai
